--- src/types/reports.ts
@@ -1,6 +1,0 @@
-export type Admin = {
-  type: string;
-  title: string;
-  target_id: string;
-  reported_user_id: string;
-};
